feat(history)!: remove bundled kip-plugin server-side provider
The SK History API is now served by external providers (InfluxDB via
signalk-to-influxdb2, or signalk-parquet); the plugin's home-baked
node:sqlite history capture is retired. After the image-asset routes were
removed, the plugin was a pure history provider, so retiring that role means
removing it wholesale.

- Delete the kip-plugin/ directory.
- Relocate the shared kip-series-contract into core/contracts (it was
  re-exported from the plugin; the kept client History-API consumers still
  need it).
- Drop plugin wiring from package.json (main, files, plugin keyword,
  enabled-by-default, build:plugin/test:plugin/build:all scripts) and the
  plugin-node-test / plugin-build CI jobs.
- Drop the plugin-only Node >=22.5 (node:sqlite) note and build:all from
  docs and .gitignore/.npmignore.

The client-side consumers (HistoryApiClientService, KipSeriesApiClientService,
HistoryToChartMapperService, HistoryChartStreamService) are kept; charts
degrade to the HISTORY_UNAVAILABLE empty state when no provider is present.

Closes #176
