+++ .npmignore
@@ -1,7 +1,5 @@
 # Manual project excludes
 /src
-/kip-plugin
-/plugin-config-data
 .angular/
 e2e/
 svg-templates/
