+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Cross-origin token sign-in: the Signal K server provides no token-refresh endpoint, so the stored password is no longer re-sent to renew a session. The session token still persists across browser reloads until it expires; when it expires you are prompted to sign in again.
 * Skip now stores its configuration in its own namespace, separate from KIP's, on both the Signal K server (applicationData appid `skip` instead of `kip`) and in browser storage (`skip.`-prefixed keys). On a server or browser that had also run KIP, Skip previously loaded KIP's dashboards and settings; it now starts from its own defaults. This is a **one-time reset with no migration** — your existing KIP config is left untouched but is not imported, so re-create or re-import your Skip configuration after this upgrade.
 * The Image widget and the server-side image asset service have been removed. Dashboards that used an Image widget will show a "failed to load" placeholder in that tile (one-time; no data migration). In-app image asset management is planned to return via a future standalone `sk-media` Signal K plugin.
+* Skip's bundled server-side history provider (the `node:sqlite`-backed Signal K plugin) has been removed, along with its "Widget Historical Data" provider selection in Settings → Display. History charts now rely on an external Signal K History API provider (e.g. InfluxDB via `signalk-to-influxdb2`, or `signalk-parquet`); where no provider is present, charts show an empty state.
 
 # v4.8.0
 ## New Features
