--- kip-plugin/tsconfig.plugin.json
@@ -1,14 +1,0 @@
-/* To learn more about Typescript configuration file: https://www.typescriptlang.org/docs/handbook/tsconfig-json.html. */
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../plugin",
-    "types": [],
-    "resolveJsonModule": true,
-    "sourceMap": false,
-    "lib": ["es2020", "es2021.intl", "DOM", "DOM.Iterable"],
-    "module": "commonjs",
-    "moduleResolution": "node"
-  },
-  "include": ["./src/**/*.ts"]
-}
